port: re-port imsg to the vendored (new) OpenBSD API; bump to 3.7.6
The imsg subsystem was ported from an older imsg generation than the tmux
this repo vendors (next-3.7). That stale port used a direct write-queue in
`struct msgbuf` and marked stack buffers with `max == 0`, whereas the vendored
imsg-buffer.c uses an `ibufqueue`, a reader-callback framing model, and the
`fd == IBUF_FD_MARK_ON_STACK` sentinel. The mismatch was the source of a
use-after-free: an ibuf freed while still linked in the send queue left a
dangling node that later crashed msgbuf_write with SIGBUS on a wild address,
plus a matching child SIGABRT from the max==0 guards.

Rewrite src/ported/compat/imsg.rs and imsg_buffer.rs to mirror the current
vendored imsg.c / imsg-buffer.c: imsgbuf owns a heap *msgbuf; msgbuf owns the
write queue and a reader-callback-driven read side (ibufqueue, rbufs, rpmsg,
imsg_parse_hdr); stack detection via fd == IBUF_FD_MARK_ON_STACK; imsg_hdr len
is u32; and a latent double-free in imsg_compose_ibuf is fixed. Update proc.rs
peer dispatch and server_client.rs header arithmetic to the new API.

Validated: 32 imsg unit tests (socketpair round-trip, fd passing, multi-message
framing), full lib suite, live client/server end-to-end, and parity 1097/1097
against vendored tmux next-3.7.

Retains the crash diagnostics from hunting this: fatal/fatalx and a
SIGSEGV/SIGBUS/SIGABRT/SIGILL/SIGFPE handler dump a backtrace to
~/.ztmux/server-crash-<pid>.txt; debug logs move to ~/.ztmux with size-based
rotation; scripts/memcheck.sh runs the server under Guard Malloc or ASan.

+++ src/ported/server_client.rs
@@ -2906,7 +2906,7 @@ pub unsafe fn server_client_dispatch(imsg: *mut imsg, arg: *mut c_void) {
             return;
         }
 
-        let datalen = (*imsg).hdr.len - IMSG_HEADER_SIZE as u16;
+        let datalen = ((*imsg).hdr.len - IMSG_HEADER_SIZE as u32) as u16;
 
         match msgtype::try_from((*imsg).hdr.type_).expect("unexpected msgtype") {
             msgtype::MSG_IDENTIFY_CLIENTPID
@@ -3121,7 +3121,7 @@ pub unsafe fn server_client_dispatch_identify(c: *mut client, imsg: *mut imsg) {
         }
 
         let data = (*imsg).data;
-        let datalen = (*imsg).hdr.len - IMSG_HEADER_SIZE as u16;
+        let datalen = ((*imsg).hdr.len - IMSG_HEADER_SIZE as u32) as u16;
 
         match msgtype::try_from((*imsg).hdr.type_).expect("unexpectd msgtype") {
             msgtype::MSG_IDENTIFY_FEATURES => {
